중고거래 상세 화면 추가

+++ danngn-back/src/main/java/com/clone/danngnback/repository/PdProductRepository.java
@@ -4,5 +4,5 @@ import com.clone.danngnback.dto.PdProduct;
 import com.clone.danngnback.dto.Test;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PdProductRepository extends JpaRepository<PdProduct, Long> {
+public interface PdProductRepository extends JpaRepository<PdProduct, Integer> {
 }
